FILE SPLITTING teamatch.py
